Refactor component functionality
The component was refactored to control state and manage props for
WeatherCurrent, WeatherForecast, and WeatherModal. The component now
determines the correct fetch URLs based on userInput and units.
Geolocation was removed as a method for fetching user's location, for
simplicity. Before this commit, I was unsure how to manage fetched data,
then I discovered custom hooks, useAxios.