Fix issue causing unmounting error
The useEffect hook calls a function getClockData that sets the current
date and time state, which is called every second. When the component is
unMounted the function still runs but (as i've learned) can be stopped
by setting getClockData to a variable and returning a function,
clearInterval(variable). This halts execution and returns a fixed bug.