Add bob ross data to component files
Originally, I used a url that would return a random painting. I wanted
to find a solution to use a JSON object that would allow me to display
better images and additional information. I brought in this data file
until I can setup a Firestore database to fetch the data from. Or leave
it, as its only ~250 KB. Not sure what to do...